refactor: separate binding adapter by view #74
