+++ presentation/src/main/java/com/kdjj/presentation/view/bindingadapter/ImageViewBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.kdjj.presentation.view.adapter
+package com.kdjj.presentation.view.bindingadapter
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
@@ -6,12 +6,6 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.kdjj.presentation.R
-
-@Suppress("UNCHECKED_CAST")
-@BindingAdapter("app:submitList")
-fun <T, VH : RecyclerView.ViewHolder> RecyclerView.submitList(list: List<T>?) {
-    (adapter as? ListAdapter<T, VH>)?.submitList(list)
-}
 
 @BindingAdapter("app:loadImage", "app:defaultImage", requireAll = false)
 fun ImageView.loadImage(src: String?, defaultResId: Int?) {
