+++ presentation/src/main/java/com/kdjj/presentation/view/bindingadapter/RecyclerViewBindingAdapter.kt
@@ -1,0 +1,11 @@
+package com.kdjj.presentation.view.bindingadapter
+
+import androidx.databinding.BindingAdapter
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+
+@Suppress("UNCHECKED_CAST")
+@BindingAdapter("app:submitList")
+fun <T, VH : RecyclerView.ViewHolder> RecyclerView.submitList(list: List<T>?) {
+    (adapter as? ListAdapter<T, VH>)?.submitList(list)
+}
